IOS-3846: Localization, cleanup and storing accepted requirements
