+++ Tangem/Resources/Localizations/de.lproj/Localizable.strings
@@ -491,6 +491,7 @@
 "wallet_currency_subtitle" = "%@ network";
 "wallet_error_no_account" = "Das Konto ist nicht erstellt";
 "wallet_network_group_title" = "%s network";
+"wallet_notification_address_copied" = "Die Adresse wurde erfolgreich kopiert";
 "wallet_pending_tx_receiving" = "Erhalten";
 "wallet_pending_tx_receiving_address_format" = "von %@";
 "wallet_pending_tx_sending" = "Senden";
