+++ Tangem/Resources/Localizations/en.lproj/Localizable.strings
@@ -491,6 +491,7 @@
 "wallet_currency_subtitle" = "%@ network";
 "wallet_error_no_account" = "Account is not created";
 "wallet_network_group_title" = "%s network";
+"wallet_notification_address_copied" = "Address was copied to clipboard";
 "wallet_pending_tx_receiving" = "Receiving";
 "wallet_pending_tx_receiving_address_format" = "from %@";
 "wallet_pending_tx_sending" = "Sending";
