+++ Tangem/Resources/Localizations/fr.lproj/Localizable.strings
@@ -491,6 +491,7 @@
 "wallet_currency_subtitle" = "%@ network";
 "wallet_error_no_account" = "Compte n'est pas créé";
 "wallet_network_group_title" = "%s network";
+"wallet_notification_address_copied" = "L'adresse a été copiée avec succès";
 "wallet_pending_tx_receiving" = "Réception";
 "wallet_pending_tx_receiving_address_format" = "de %@";
 "wallet_pending_tx_sending" = "Envoi";
