+++ Tangem/Resources/Localizations/it.lproj/Localizable.strings
@@ -491,6 +491,7 @@
 "wallet_currency_subtitle" = "%@ network";
 "wallet_error_no_account" = "Conto non creato";
 "wallet_network_group_title" = "%s network";
+"wallet_notification_address_copied" = "L'indirizzo è stato copiato con successo";
 "wallet_pending_tx_receiving" = "Riceviu";
 "wallet_pending_tx_receiving_address_format" = "da %@";
 "wallet_pending_tx_sending" = "Inviau";
