+++ Tangem/Resources/Localizations/zh-Hant.lproj/Localizable.strings
@@ -491,6 +491,7 @@
 "wallet_currency_subtitle" = "%@ 網路";
 "wallet_error_no_account" = "帳號尚未被創造";
 "wallet_network_group_title" = "%s network";
+"wallet_notification_address_copied" = "地址已復製到剪貼板";
 "wallet_pending_tx_receiving" = "接收中";
 "wallet_pending_tx_receiving_address_format" = "來自%@";
 "wallet_pending_tx_sending" = "發送中";
